skip_genshin/requirements.txt
keyboard==0.13.5
MouseInfo==0.1.3
mss==10.1.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pynput==1.8.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==311
six==1.17.0